Some work towards making the login screen closer to completion (#52)
* Remove duplicate "add to existing" checkbox in login view

That messes some things up. If it was a stub for some other
unimplemented controls, I don't think it should have been left in
like that, and I'll implement actual controls later.

* Rework login screen to not use grid entirely

Needed to avoid weirdness with non-text-field inputs below

* Refactor the "add to existing" control

I put the dropdown on the same line (felt weird to have it on two lines),
removed dependency on the Dropdown component, and made it show (greyed out)
even if the checkbox is off.

* Implement the "is a farm" checkbox

Added some changes to conform closer to the original design specification.

* Implement visuals for the URL input field / area

* Add visuals for the "save as" field

* Add non-functional "submit" button

* Implement layout as originally specified

* Implement spacing in a manner compatible with the vertical line

* Use a better approach to spacing

* Change overflow scrollbars to appear in the active tab

Before they used to appear for the app window itself, which was
rather disruptive.

* Improve multi-URL textarea layout

Letting it resize vertically caused some rather weird issues. I made
it fall back to the scrolling behavior.

* Make the mage unselectable

* Add horizontal space for extra-wide screens