Implement settings saving in login form
Tested to save the settings when creating a new wiki or a wiki farm.
Partially tested attachment to existing wikis (that feature simply does
not work yet), doesn't crash at least.

Accompanied by a spellbook rewrite. This rewrite was motivated by that:
 - the old spellbook exported absolutely everything;
 - it used field-like syntax with some of its functions, which was
   not as readable as actual functions;
 - it contained multiple monolithic functions that coupled a lot of
   functionality, including three HTTP requests for adding a new wiki,
   of which at least one was performed even if likely redundant;
 - it contained some issues, such as silently overwriting keytar on
   every settings save, and spammed some errors into the console;
 - it had no documentation for anything, which would have helped given
   its use of nontrivial structures and counter-intuitive identifiers...

I also "sneaked in" a change to how the "overwrite keytar" feature works.
Instead of being completely silent, it now prompts the user whether they
want to delete the passwords (not by default), and it only shows the
prompt if there are passwords to delete. That latter one is important
because otherwise every new user would trigger the prompt without there
being any passwords to delete.