More login screen work (#55)
* Extract login form validation into its own method

Now it's separate from saving. Feels a bit cleaner.

* Change the bad remote error to include the channel name

For easier debugging. Also changed the error text.

* Use rest parameters in window.api.remote calls

* Implement settings saving in login form

Tested to save the settings when creating a new wiki or a wiki farm.
Partially tested attachment to existing wikis (that feature simply does
not work yet), doesn't crash at least.

Accompanied by a spellbook rewrite. This rewrite was motivated by that:
 - the old spellbook exported absolutely everything;
 - it used field-like syntax with some of its functions, which was
   not as readable as actual functions;
 - it contained multiple monolithic functions that coupled a lot of
   functionality, including three HTTP requests for adding a new wiki,
   of which at least one was performed even if likely redundant;
 - it contained some issues, such as silently overwriting keytar on
   every settings save, and spammed some errors into the console;
 - it had no documentation for anything, which would have helped given
   its use of nontrivial structures and counter-intuitive identifiers...

I also "sneaked in" a change to how the "overwrite keytar" feature works.
Instead of being completely silent, it now prompts the user whether they
want to delete the passwords (not by default), and it only shows the
prompt if there are passwords to delete. That latter one is important
because otherwise every new user would trigger the prompt without there
being any passwords to delete.

* Remove stray comment

Forgot to remove it with the previous commit.

* Block the save button when saving

Not necessarily critical, but might prevent some double-saving.

* Notify the user of successful login data save

... and reset the form.

* Hook up the auth system list in the login form to the actual settings

* Extract settings updating into their own function

Cleaner promise chains + less redundancy in calling the success handler

* Remove extra newline

* Remove no longer used variable

* Overload the spellbook with documentation and other comments

After all, lack of documentation was one of my issues with the original
spellbook.