Task screen work (#58)
* Upgrade Tailwind to 2.1.0

Needed for filter and backdrop filter support

* Implement a stub for the page list generator modal

This is, of course, heavily experimental. The two test generators,
"Numbers" and "Letters", are there for testing and because I don't have
any actual API code yet.

* Implement basic task-adding functionality

Unfortunately, involves a LOT of duplication. It's really not DRY.
Not sure how I'd fix it though. Also changed some things with the
generator modal for consistency with the new add-tasks modal.

* Implement a basic task runner

Does nothing but log messages to the dev tools right now. Useful
for testing, but not for anything practical.

* Change all SVG icons to be unselectable

Behaves more like typical program user interfaces.

* Fix login window checkbox inputs

They used to have an unclickable area between the label and the checkbox.
Trivial fix, might save some frustration for those unlucky enough to
drive their cursor to that area. Only relation with other commits is that
I learned how to "better" fix this while making task window modals.