Rework the login screen
This might make it look a bit less weird.
Known minor issue (couldn't resolve): if the help page title wraps, and
the page used is non-default (has the home icon), the spacing will look
as if justify-between, not justify-evenly. (Effectively, when there's a
line wrap in the title element, the flex width doesn't recalculate.)
Not sure how to fix it.