More minor changes (#70)
* Bump dependencies

Nothing appears to be broken.

* Refactor the login form to remove both checkboxes

Advantages:
 - No longer possible to have a problematic case when the user selects
   "add to existing", but there are no wiki farms or standalone wikis
   to add new credentials to.
 - Simplifies some internal code a bit. Also changed some variable names
   to be clearer.
Disadvantages:
 - May not make it clear that it's needed to enter that dropdown if one wants
   to set up a new standalone wiki, rather than a new wiki farm. I'm looking
   to address it in the next commit(s). Thanks to River for pointing that
   out.

* Rework the login screen

This might make it look a bit less weird.
Known minor issue (couldn't resolve): if the help page title wraps, and
the page used is non-default (has the home icon), the spacing will look
as if justify-between, not justify-evenly. (Effectively, when there's a
line wrap in the title element, the flex width doesn't recalculate.)
Not sure how to fix it.

* Improve light theme contrast for special options

Forgot that minor detail.